requirements.txt
fastapi==0.123.9
hypercorn==0.14.4
elevenlabs==1.94.1
python-dotenv==1.0.1
